requirements.txt
Flask>=3.0,<4.0
flask-sock>=0.7,<1.0
python-dotenv>=1.0,<2.0
pyserial>=3.5,<4.0
psutil>=5.9,<7.0
RPi.GPIO>=0.7; platform_machine == "aarch64" or platform_machine == "armv7l"
adafruit-circuitpython-ina3221; platform_machine == "aarch64" or platform_machine == "armv7l"
adafruit-circuitpython-bme280; platform_machine == "aarch64" or platform_machine == "armv7l"
adafruit-circuitpython-lis2dw12; platform_machine == "aarch64" or platform_machine == "armv7l"
smbus2; platform_machine == "aarch64" or platform_machine == "armv7l"
